ExFleetYards.Repo.Account.User (Fleetyards v0.1.0)

User

Types

t()

@type t() :: %ExFleetYards.Repo.Account.User{
  __meta__: Ecto.Schema.Metadata.t(),
  avatar: String.t() | nil,
  confirmation_sent_at: NaiveDateTime.t() | nil,
  confirmation_token: String.t() | nil,
  confirmed_at: NaiveDateTime.t() | nil,
  consumed_timestep: integer() | nil,
  created_at: DateTime.t() | nil,
  current_sign_in_at: NaiveDateTime.t() | nil,
  current_sign_in_ip: String.t() | nil,
  discord: String.t() | nil,
  email: String.t() | nil,
  encrypted_otp_secret: String.t() | nil,
  encrypted_otp_secret_iv: String.t() | nil,
  encrypted_otp_secret_salt: String.t() | nil,
  encrypted_password: String.t() | nil,
  failed_attempts: integer() | nil,
  guilded: String.t() | nil,
  hangar_updated_at: NaiveDateTime.t() | nil,
  homepage: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  last_sign_in_at: NaiveDateTime.t() | nil,
  locale: String.t() | nil,
  locked_at: NaiveDateTime.t() | nil,
  normalized_email: String.t() | nil,
  normalized_username: String.t() | nil,
  otp_backup_codes: [String.t()] | nil,
  otp_required_for_login: boolean() | nil,
  password: String.t() | nil,
  public_hangar: boolean() | nil,
  public_hangar_loaners: boolean() | nil,
  remember_created_at: NaiveDateTime.t() | nil,
  reset_password_sent_at: NaiveDateTime.t() | nil,
  reset_password_token: String.t() | nil,
  rsi_handle: String.t() | nil,
  sale_notify: boolean() | nil,
  sign_in_count: integer() | nil,
  tracking: boolean() | nil,
  twitch: String.t() | nil,
  unconfirmed_email: String.t() | nil,
  unlock_token: String.t() | nil,
  updated_at: DateTime.t() | nil,
  username: String.t() | nil,
  vehicles: Ecto.Schema.has_many(ExFleetYards.Repo.Account.Vehicle.t()),
  youtube: String.t() | nil
}

Functions

confirm_changeset(user, attrs)

info_changeset(user, attrs)

registration_changeset(user \\ %__MODULE__{}, attrs)

valid_password?(arg1, password)

Verifies the password. If there is no user or the user doesn't have a password, we call Bcrypt.no_user_verify/0 to avoid timing attacks.

validate_current_password(changeset, password)

Validates the current password otherwise adds an error to the changeset.