ExFleetYards.Repo.Account.UserToken (Fleetyards v0.1.0)

User Token

Functions

build_api_token(user, scopes)

build_auth_token(user)

build_hashed_token(user, context \\ "api")

Generates a token that will be stored in a signed place, such as session or cookie. As they are signed, those tokens do not need to be hashed.

changeset(token \\ %__MODULE__{}, attrs)

create_confirm_token(user)

scopes()

token_and_context_query(token, context)

Returns the given token with the given context.

transform_scope(arg)

transform_scopes(scopes)

user_and_contexts_query(user, context)

Gets all tokens for the given user for the given contexts.

user_and_id_query(user, id)

verify_hashed_token(token, context \\ "api")

verify_token(token, context)

verify_token(token, context, days)