ExFleetYards.Repo.Changeset (Fleetyards v0.1.0)

Changeset validators for ExFleetYards

change_url_or_path(changeset, key, host)

@spec change_url_or_path(Ecto.Changeset.t(), atom(), String.t()) :: Ecto.Changeset.t()

Change given change url to a full url if only path is provided

validate_discord_server(changeset, key \\ :discord)

@spec validate_discord_server(Ecto.Changeset.t(), atom()) :: Ecto.Changeset.t()

Validate discord server invite link

validate_rsi_handle(changeset, key \\ :rsi_handle)

@spec validate_rsi_handle(Ecto.Changeset.t(), atom()) :: Ecto.Changeset.t()

Validate rsi handle

validate_twitch(changeset, key \\ :twitch)

@spec validate_twitch(Ecto.Changeset.t(), atom()) :: Ecto.Changeset.t()

Validate twitch link

validate_url_host(changeset, key, host, opts \\ [])

@spec validate_url_host(
  Ecto.Changeset.t(),
  atom(),
  String.t() | [String.t()],
  keyword()
) ::
  Ecto.Changeset.t()

Validate given change url is on a given host.

validate_youtube(changeset, key \\ :youtube)

@spec validate_youtube(Ecto.Changeset.t(), atom()) :: Ecto.Changeset.t()

Validate youtube link