ExFleetYardsApi.Auth (Fleetyards v0.1.0)

Auth helpers

Functions

check_fleet(token, fleet, rank \\ :member)

check_fleet_scope(token, fleet, scopes, rank \\ :member)

@spec check_fleet_scope(
  Plug.Conn.t() | ExFleetYards.Repo.Account.UserToken.t(),
  ExFleetYards.Repo.Fleet.t() | binary(),
  [binary()] | binary(),
  atom()
) :: ExFleetYards.Repo.Fleet.t()

Check if the token has the required scopes for the fleet

check_fleet_scope_or_public(conn, fleet, scopes)

@spec check_fleet_scope_or_public(
  Plug.Conn.t() | ExFleetYards.Repo.Account.UserToken.t(),
  ExFleetYards.Repo.Fleet.t() | binary(),
  [binary()] | binary()
) :: {boolean(), ExFleetYards.Repo.Fleet.t()}

Check that the token has access to the fleet, or return the public fleet

check_scope(access, token)

check_scopes(token, scopes)

fetch_api_token(conn, opts)

required_api_scope(conn, scopes)