ExFleetYardsApi.JsonHelpers (Fleetyards v0.1.0)

Helper functions for rendering JSON.

add_if(map, key, data, cond)

@spec add_if(map(), atom(), any(), boolean()) :: map()

Add if condition is true

filter_null(map, required \\ [])

@spec filter_null(map(), [atom()] | atom()) :: map()

Filter out nil values from a map, unless they are required.

Required fields are specified by a list of atoms, or a schema module.

render_meta(page)

@spec render_meta(Chunkr.Page.t()) :: map()

Render a cursor-based pagination meta object.

render_page(page, module, render_one, params \\ %{})

Render a page of data using a module and function. The function has to take in the argument data and return a map.

render_timestamp(data)

@spec render_timestamp(DateTime.t() | nil) :: String.t() | nil

Render a timestamp as an ISO8601 string.

render_timestamp(map, key, data)

render_timestamps(map, data)

@spec render_timestamps(map(), %{
  created_at: DateTime.t() | nil,
  updated_at: DateTime.t() | nil
}) ::
  map()

Renders created_at and updated_at timestamps as ISO8601 strings.