ExFleetYardsApi.Plugs.Authorization behaviour (Fleetyards v0.1.0)

This module provides plugs for authorization and authentication.

Plugs

require_authenticated/2

Ensures that the user is authenticated by checking for a valid bearer token in the authorization header.

It assigns the current token and user to the connection.

authorize/2

Ensures that the user has the required OAuth2 scopes.

Takes a list of required scopes as a parameter.

authorize(t, t)

@callback authorize(Plug.Conn.t(), Keyword.t()) :: Plug.Conn.t()

Ensures that the user has the required OAuth2 scopes.

Takes a list of required scopes as a parameter.

require_authenticated(t, t)

@callback require_authenticated(Plug.Conn.t(), Keyword.t()) :: Plug.Conn.t()

Ensures that the user is authenticated by checking for a valid bearer token in the authorization header.

It assigns the current token and user to the connection.

authorize(conn, scopes)

Ensures that the user has the required OAuth2 scopes.

Takes a list of required scopes as a parameter.

require_authenticated(conn, scopes)

Ensures that the user is authenticated by checking for a valid bearer token in the authorization header.

It assigns the current token and user to the connection.