ExFleetYardsApi.Router.Helpers (Fleetyards v0.1.0)

Module with named helpers generated from ExFleetYardsApi.Router.

Functions

celestial_object_path(conn_or_endpoint, action)

celestial_object_path(conn_or_endpoint, action, params)

celestial_object_path(conn_or_endpoint, action, id, params)

celestial_object_url(conn_or_endpoint, action)

celestial_object_url(conn_or_endpoint, action, params)

celestial_object_url(conn_or_endpoint, action, id, params)

component_path(conn_or_endpoint, action)

component_path(conn_or_endpoint, action, params)

component_path(conn_or_endpoint, action, id, params)

component_url(conn_or_endpoint, action)

component_url(conn_or_endpoint, action, params)

component_url(conn_or_endpoint, action, id, params)

fleet_invite_path(conn_or_endpoint, action, token)

fleet_invite_path(conn_or_endpoint, action, token, params)

fleet_invite_path(conn_or_endpoint, action, slug, user, params)

fleet_invite_url(conn_or_endpoint, action, token)

fleet_invite_url(conn_or_endpoint, action, token, params)

fleet_invite_url(conn_or_endpoint, action, slug, user, params)

fleet_path(conn_or_endpoint, action)

fleet_path(conn_or_endpoint, action, params)

fleet_path(conn_or_endpoint, action, slug, params)

fleet_url(conn_or_endpoint, action)

fleet_url(conn_or_endpoint, action, params)

fleet_url(conn_or_endpoint, action, slug, params)

manufacturer_path(conn_or_endpoint, action)

manufacturer_path(conn_or_endpoint, action, params)

manufacturer_path(conn_or_endpoint, action, id, params)

manufacturer_url(conn_or_endpoint, action)

manufacturer_url(conn_or_endpoint, action, params)

manufacturer_url(conn_or_endpoint, action, id, params)

model_path(conn_or_endpoint, action)

model_path(conn_or_endpoint, action, params)

model_path(conn_or_endpoint, action, id, params)

model_url(conn_or_endpoint, action)

model_url(conn_or_endpoint, action, params)

model_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

render_spec_path(conn_or_endpoint, action)

render_spec_path(conn_or_endpoint, action, params)

render_spec_url(conn_or_endpoint, action)

render_spec_url(conn_or_endpoint, action, params)

roadmap_path(conn_or_endpoint, action)

roadmap_path(conn_or_endpoint, action, params)

roadmap_path(conn_or_endpoint, action, id, params)

roadmap_url(conn_or_endpoint, action)

roadmap_url(conn_or_endpoint, action, params)

roadmap_url(conn_or_endpoint, action, id, params)

star_system_path(conn_or_endpoint, action)

star_system_path(conn_or_endpoint, action, params)

star_system_path(conn_or_endpoint, action, id, params)

star_system_url(conn_or_endpoint, action)

star_system_url(conn_or_endpoint, action, params)

star_system_url(conn_or_endpoint, action, id, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

station_path(conn_or_endpoint, action)

station_path(conn_or_endpoint, action, params)

station_path(conn_or_endpoint, action, id, params)

station_path(conn_or_endpoint, action, id, shop, params)

station_url(conn_or_endpoint, action)

station_url(conn_or_endpoint, action, params)

station_url(conn_or_endpoint, action, id, params)

station_url(conn_or_endpoint, action, id, shop, params)

swagger_ui_path(conn_or_endpoint, action)

swagger_ui_path(conn_or_endpoint, action, params)

swagger_ui_url(conn_or_endpoint, action)

swagger_ui_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_hangar_path(conn_or_endpoint, action)

user_hangar_path(conn_or_endpoint, action, params)

user_hangar_path(conn_or_endpoint, action, model, params)

user_hangar_url(conn_or_endpoint, action)

user_hangar_url(conn_or_endpoint, action, params)

user_hangar_url(conn_or_endpoint, action, model, params)

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, token, params)

user_session_path(conn_or_endpoint, action)

user_session_path(conn_or_endpoint, action, params)

user_session_path(conn_or_endpoint, action, id, params)

user_session_url(conn_or_endpoint, action)

user_session_url(conn_or_endpoint, action, params)

user_session_url(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, token, params)

version_path(conn_or_endpoint, action)

version_path(conn_or_endpoint, action, params)

version_url(conn_or_endpoint, action)

version_url(conn_or_endpoint, action, params)