ExFleetYardsAuth.Router.Helpers (Fleetyards v0.1.0)

Module with named helpers generated from ExFleetYardsAuth.Router.

Functions

authorize_path(conn_or_endpoint, action)

authorize_path(conn_or_endpoint, action, params)

authorize_url(conn_or_endpoint, action)

authorize_url(conn_or_endpoint, action, params)

introspect_path(conn_or_endpoint, action)

introspect_path(conn_or_endpoint, action, params)

introspect_url(conn_or_endpoint, action)

introspect_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

revoke_path(conn_or_endpoint, action)

revoke_path(conn_or_endpoint, action, params)

revoke_url(conn_or_endpoint, action)

revoke_url(conn_or_endpoint, action, params)

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

token_path(conn_or_endpoint, action)

token_path(conn_or_endpoint, action, params)

token_url(conn_or_endpoint, action)

token_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.